1. Check exist before modify files 2. Add tar binary provide --checkpoint option 3. Add chmod +x in init.sh

+++ ramfs_setup.sh
@@ -3,21 +3,21 @@
 # Exit if there is any error
 set -e
 
-echo "Setup GRUB ..."
-if ! test -f /etc/default/grub.original; then
-	mv /etc/default/grub /etc/default/grub.original
-fi
-cat << EOF > /etc/default/grub
-GRUB_DEFAULT=0
-GRUB_TIMEOUT=5
-GRUB_DISTRIBUTOR=\`lsb_release -i -s 2> /dev/null || echo Debian\`
-GRUB_CMDLINE_LINUX_DEFAULT=""
-GRUB_CMDLINE_LINUX="console=tty0 console=ttyS0,115200n8"
-GRUB_TERMINAL="console serial"
-GRUB_SERIAL_COMMAND="serial --speed=115200 --unit=0 --word=8 --parity=no --stop"
-GRUB_DISABLE_OS_PROBER=true
-EOF
-update-grub
+# echo "Setup GRUB ..."
+# if ! test -f /etc/default/grub.original; then
+# 	mv /etc/default/grub /etc/default/grub.original
+# fi
+# cat << EOF > /etc/default/grub
+# GRUB_DEFAULT=0
+# GRUB_TIMEOUT=5
+# GRUB_DISTRIBUTOR=\`lsb_release -i -s 2> /dev/null || echo Debian\`
+# GRUB_CMDLINE_LINUX_DEFAULT=""
+# GRUB_CMDLINE_LINUX="console=tty0 console=ttyS0,115200n8"
+# GRUB_TERMINAL="console serial"
+# GRUB_SERIAL_COMMAND="serial --speed=115200 --unit=0 --word=8 --parity=no --stop"
+# GRUB_DISABLE_OS_PROBER=true
+# EOF
+# update-grub
 
 echo "Install additional software ..."
 apt update
@@ -25,19 +25,47 @@ apt install htop screen wget tcpdump dosfstools exfat-utils mdadm lvm2 linux-cpu
 apt clean
 
 echo "Customizing initramfs script ..."
-cp /usr/share/initramfs-tools/scripts/local /usr/share/initramfs-tools/scripts/local.original
+if ! test -f /usr/share/initramfs-tools/scripts/local.original; then
+     cp /usr/share/initramfs-tools/scripts/local /usr/share/initramfs-tools/scripts/local.original
+fi
 cp local.ramfs /usr/share/initramfs-tools/scripts/local.ramfs
 
 echo "Adding additional kernel module for FAT and exFAT ..."
-cp /etc/initramfs-tools/modules /etc/initramfs-tools/modules.original
-cp /etc/initramfs-tools/modules /etc/initramfs-tools/modules.ramfs
-ls /lib/modules/`uname -r`/kernel/fs/fat/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
-ls /lib/modules/`uname -r`/kernel/fs/exfat/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
-ls /lib/modules/`uname -r`/kernel/fs/nls/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
+if ! test -f /etc/initramfs-tools/modules.original; then
+     cp /etc/initramfs-tools/modules /etc/initramfs-tools/modules.original
+fi
+if ! test -f /etc/initramfs-tools/modules.ramfs; then
+     cp /etc/initramfs-tools/modules /etc/initramfs-tools/modules.ramfs
+     ls /lib/modules/`uname -r`/kernel/fs/fat/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
+     ls /lib/modules/`uname -r`/kernel/fs/exfat/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
+     ls /lib/modules/`uname -r`/kernel/fs/nls/ | cut -f1 -d '.' | tee -a /etc/initramfs-tools/modules.ramfs
+fi
+
+echo "Adding tar binary to initramfs"
+cat << EOF > /usr/share/initramfs-tools/hooks/tar
+#!/bin/sh
+PREREQ=""
+prereqs()
+{
+     echo "\$PREREQ"
+}
+case \$1 in
+prereqs)
+     prereqs
+     exit 0
+     ;;
+esac
+. /usr/share/initramfs-tools/hook-functions
+rm -f \${DESTDIR}/bin/tar
+copy_exec /usr/bin/tar /bin/tar
+EOF
+chmod +x /usr/share/initramfs-tools/hooks/tar
 
 echo "Customizing fstab ..."
-cp /etc/fstab /etc/fstab.original
-echo 'none /   tmpfs size=95% 0 0' | tee -a /etc/fstab.ramfs
+if ! test -f /etc/fstab.original; then
+     cp /etc/fstab /etc/fstab.original
+fi
+echo 'none /   tmpfs size=95% 0 0' | tee /etc/fstab.ramfs
 
 echo "Setup startup.service ..."
 cat << EOF > /etc/systemd/system/startup.service
@@ -77,13 +105,15 @@ echo "Generating initramfs script ..."
 cat << EOF > /bootfiles/init.sh
 #!/bin/sh
 
-echo "Copying start.sh ..."
-cp /mount/start.sh .
 echo "Copying rootfs.tar.gz ..."
-cp /mount/rootfs.tar.gz .
+cp /mount/EFI/rootfs.tar.gz .
+echo "Copying start.sh ..."
+cp /mount/EFI/start.sh .
+chmod +x start.sh
+echo "Unmount boot device ..."
 umount /mount
-echo "Extracting from rootfs.tar.gz ..."
-tar zxvf rootfs.tar.gz
+echo -n "Extracting from rootfs.tar.gz ..."
+tar zxf rootfs.tar.gz --checkpoint=.5000
 rm rootfs.tar.gz
 EOF
